requirement.txt
spacy >= 3.0
genai >= 2.0
python >= 3.11
imblearn
python-dot_env >= 1.0
google.transit
pandas

